requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.5
djangorestframework==3.12.4
idna==3.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
